SpinAPI deprecated the use of pb_get_error.
Although the function call is still present in the C headers, there's no actual definition to it anymore
and the spinAPI will just return an empty string if this function is called. Its functionality has been
replaced by SpinCore with pb_status_message(), which now not only tells you the state of the board (running,
reset, etc), but also whatever error state it is in or what error it encountered (such as a failure to
communicate). Unfortunately, the python wrapper being used to call these functions still references the
deprecated function pb_get_error, so whenever the PB encounters an error the blacs worker passes through
the empty error string, making troubleshooting difficult. Therefore a function override was added to the
device file for the PulseBlaster tp call pb_status_message instead when pb_get_error is invoked.