NI Device recovery due to bandwidth issues and puleblaster reconnect
Modifications to the pulseblaster_no_dds file were added to attempt the yield function again on a timeout exception during communication between the worker and its parent. The pulseblaster would randomly lock up due to a hangup during communication between workers and their parent threads. So a timeout option was added in an earlier commit. Now if a timeout is encountered during running of status_monitor the function will make another attempt before throwing a final timeout exception.

Additionally, parameters were added to the labscript device file of the NI boards, and its worker was modified, to support a work around for an unfortunate bug where the NI devices enter an error state that's unrecoverable after encountering a bandwidth error, Error -200946. This would force the user to physically disconnect the USB cables from the NI device and reconnect them to bring them back. To circumvent this, if such an error is encountered and the FORCE_RESTART_USB flag is turned on, BLACS will attempt to restart the USB bus the NI devices are connected to via a console command invoking the windows utility devcon and have windows re-initialize all connected devices in transition to manual. Essentially simulating the act of physically disconnecting and reconnecting the cable. After this is complete the worker will reset the device and task, and continue to the next shot after noting the current shot has failed. This is a very specific fix that will only work on Windows machines if the user goes out of their way to set up what's needed and turn it on. If the flag is off blacs will simply note that the error has occured and attempt to restart the device through the NI-DAQ driver. To use this “fix” you must:
1)Go to the device manager, expand Universal Serial Bus Controllers

2)Check the properties of each hub and check the power tab. The correct hub will have the NI devices listed.

3)Go to the details tab and change the drop down list to “Device Instance Path”

4)Write down the value between the & and second \. It could also be between two \ \. So, for example, if it said USB\VID_0424&PID_2514\5&1E7355BF&0&5 you would write down “PID_2514”. If it said USB\ROOT_HUB10\5&1E7355BF&0&5 you would write down “ROOT_HUB10”.

5)Do this for every NI device on a separate HUB. All NI devices on the same hub will be reset.

6)In the connection table for each NI device entry add a new field called USB_bus_name and set it equal to the path id you wrote down. Also add an entry called FORCE_RESTART_USB and set it equal to True For example:
NI_USB_6229(

    name='ni_usb_6229_0',

    parent_device=pulseblaster_0_ni_usb_0_clock,

    MAX_name='Dev4',

    clock_terminal='/Dev4/PFI0',

    USB_bus_name='PID_2514',

    FORCE_RESTART_USB=True

)

7)Install Windows Visual Studio. When installing select “Individual Components”. Search for “Windows 10 SDK” and select the newest revision.

8)Install Windows Driver Kit

9)Run BLACS as an Administrator