Added ability to specify timeout value and re-attempt worker/parent communication
Added functionality to Pulseblaster device files to set a timeout value as a workaround for a bug in which, after a long period of time, the Pulseblaster will no longer respond for an extended period (days) until the BLACs tab is restarted. This also occasionally creates further issues for other devices reliant on its triggering. This would occur in both manual and buffered modes. It turned out that under specific situations Labscript would hang up on the put() and get() commands during communication between a worker and its parent thread. Although I haven't quite figured out why this is happening, having a timeout value prevents this hangup from occurring and allows one to recover. Code was also added to status_monitor since this issue seems to happen the most when invoking the yield command. This code now attempts to retry the yield command if the worker times out. If the process is unable to talk again then BLACs is notified the shot has failed and the tab is restarted.