Workarounds for specific NI Errors
Modifications were made in the output worker transition_to_manual for NI boards to capture errors that do not impede the ability for an experiment to proceed and avoid putting the tab in an error state. Instead BLACs notes that the shot failed and restarts the NI device and its task so the next shot can proceed. The first two errors can occur due to noise on the line and does not necessarily require that the entire tab be put in an error state. The final error is error -200946 which in the ChipLab was causing all NI devices on the same USB bus to lock up. This error is fatal in that it requires physically unplugging all of the devices and restarting BLACs to recover from. So a catch was made to attempt to restart the NI device before this happens and note that the shot failed. Future revisions will add an option to have Windows re-initialize the bus to simulate unplugging and plugging the devices.