Integration of workaround for NI error -200946 into NI blacs tab
This is an attempt to integrate the devcon fix for NI Bandwith errors into the NI blacs tab rather than through the connection table. A checkmark box has been added to the bottom of the tab that, once checked, will enable the removal and re-initialization of the NI device when encountering error -200946. A field box will appear to allow a user to enter the device ID. Once entered, if the user has installed the necessary requirements (devcon) and entered a valid device ID, blacs will instruct windows to remove and re-initialize said device through the device manager when encountering this error. Essentially unplugging the device and plugging it back in. Once finished blacs will end the shot through normal means, note the error, and move on to the next shot file. If devcon cannot be located, or other requirements are not met, or the OS isn't a windows system, an error will be noted and the shot will be ended. Blacs will also remember the entered device ID so this only has to be entered once.

To use said feature the Windows SDK https://developer.microsoft.com/en-us/windows/downloads/windows-10-sdk/ and the Windows Development Kit https://go.microsoft.com/fwlink/?linkid=2128854 must be installed.

The windows device ID can be found in the device manager. The information can be found on the details tab for the device under "Device Instance Path". The Device ID is the value between the & and second \ characters. It could also be between two \ \ characters. So, for example, if it said USB\VID_0424&PID_2514\5&1E7355BF&0&5 you would write down “PID_2514”. If it said USB\ROOT_HUB10\5&1E7355BF&0&5 you would write down “ROOT_HUB10”.